# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "PostResponse": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "user": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "likes": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "comments": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "image": {
                        "type": "string"
                    },
                    "createdAt": {
                        "oneOf": [
                            {
                                "format": "date",
                                "type": "string"
                            },
                            {
                                "format": "date-time",
                                "type": "string"
                            }
                        ]
                    },
                    "updatedAt": {
                        "oneOf": [
                            {
                                "format": "date",
                                "type": "string"
                            },
                            {
                                "format": "date-time",
                                "type": "string"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "user",
                    "content",
                    "likes",
                    "comments",
                    "image",
                    "createdAt",
                    "updatedAt"
                ]
            },
            "CommentResponse": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "user": {
                        "type": "string"
                    },
                    "post": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "likes": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "replies": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "replyTo": {
                        "type": "string"
                    },
                    "createdAt": {
                        "oneOf": [
                            {
                                "format": "date",
                                "type": "string"
                            },
                            {
                                "format": "date-time",
                                "type": "string"
                            }
                        ]
                    },
                    "updatedAt": {
                        "oneOf": [
                            {
                                "format": "date",
                                "type": "string"
                            },
                            {
                                "format": "date-time",
                                "type": "string"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "user",
                    "post",
                    "content",
                    "likes",
                    "replies",
                    "replyTo",
                    "createdAt",
                    "updatedAt"
                ]
            },
...
            "SubmitPostRequest": {
                "properties": {
                    "content": {
                        "type": "string",
                        "maxLength": 3000,
                        "minLength": 15
                    }
                },
                "type": "object",
                "required": [
                    "content"
                ]
            },
            "CommentOnPostRequest": {
                "properties": {
                    "content": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 5
                    }
                },
                "type": "object",
                "required": [
                    "content"
                ]
            }
        }
    },
    "info": {
        "title": "SocialApp API",
        "version": "1.0.0",
        "description": "Generated with `routing-controllers-openapi`"
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/posts/": {
            "post": {
                "operationId": "PostController.submitPost",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubmitPostRequest"
                            }
                        }
                    },
                    "description": "SubmitPostRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Submit a post",
                "tags": [
                    "Post"
                ],
                "description": "\n    Submit a post with an optional image.\n    Minimum length of content is 15 characters and maximum is 3000 characters.\n    To submit an image, use Postman and send a multipart/form-data request with a file field with the key 'image'.\n    The image must be of type png, jpg or jpeg and must not exceed 1MB in size.\n    ",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "get": {
                "operationId": "PostController.getTimelinePosts",
                "parameters": [
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "lastDocumentId",
                        "schema": {
                            "pattern": "^[0-9a-fA-F]{24}$",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/PostResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get timeline posts",
                "tags": [
                    "Post"
                ],
                "description": "\n    Get a list of posts of the users that the current user is following.\n    The posts are paginated and sorted in descending order of creation date.\n    limit: limit of documents to return, default is 5.\n    lastDocumentId: id of the last document in the previous page. If not provided, it will return the first page.\n    ",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{postId}": {
            "patch": {
                "operationId": "PostController.updatePost",
                "parameters": [
                    {
                        "in": "path",
                        "name": "postId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubmitPostRequest"
                            }
                        }
                    },
                    "description": "SubmitPostRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Update a post",
                "tags": [
                    "Post"
                ],
                "description": "\n    Update the content of a post.\n    Minimum length of content is 15 characters and maximum is 3000 characters.\n    Updating the content is allowed within 1 hour of submitting the post.\n    ",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "operationId": "PostController.deletePost",
                "parameters": [
                    {
                        "in": "path",
                        "name": "postId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Delete a post",
                "tags": [
                    "Post"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{postId}/like": {
            "post": {
                "operationId": "PostController.likePost",
                "parameters": [
                    {
                        "in": "path",
                        "name": "postId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Like a post",
                "tags": [
                    "Post"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{postId}/comment": {
            "post": {
                "operationId": "PostController.commentOnPost",
                "parameters": [
                    {
                        "in": "path",
                        "name": "postId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentOnPostRequest"
                            }
                        }
                    },
                    "description": "CommentOnPostRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Comment on a post",
                "tags": [
                    "Post"
                ],
                "description": "\n    Submit a comment on a post.\n    Minimum length of content is 5 characters and maximum is 1000 characters.\n    ",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{userId}": {
            "get": {
                "operationId": "PostController.getUserPosts",
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "lastDocumentId",
                        "schema": {
                            "pattern": "^[0-9a-fA-F]{24}$",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/PostResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "summary": "Get user posts",
                "tags": [
                    "Post"
                ],
                "description": "\n    Get a list of posts of a user.\n    The posts are paginated and sorted in descending order of creation date.\n    limit: limit of documents to return, default is 5.\n    lastDocumentId: id of the last document in the previous page. If not provided, it will return the first page.\n    ",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}